requirements.txt
torch==2.2.1
onnx==1.15.0
onnxruntime==1.17.1
scikit-learn==1.4.1.post1
tqdm==4.66.2
numpy==1.26.4
pillow==10.2.0
